+++ verified_destinations/pinecone/specs.yml
