+++ verified_destinations/pinecone/tests/conftest.py
