+++ verified_destinations/qdrant/specs.yml
